:art: add static_cast to avoid implicit conversion.

+++ include/fiction/utils/math_utils.hpp
@@ -36,7 +36,7 @@ T round_to_n_decimal_places(const T number, const uint64_t n) noexcept
     if constexpr (std::is_arithmetic_v<T>)
     {
         const T factor = std::pow(10, n);
-        return std::round(number * factor) / factor;
+        return std::round(number * factor) / static_cast<double>(factor);
     }
     else if constexpr (units::traits::is_temperature_unit<T>::value || units::traits::is_length_unit<T>::value ||
                        units::traits::is_voltage_unit<T>::value || units::traits::is_energy_unit<T>::value)
